Hmm. Init panel inside constructor.
I was thinking we had to wait to call panel.port.emit, hence, needed to
queue up callbacks until it was ready. Created whenPortReady callback
inside Panel for just this reason.

However, looking through the file again, there's no reason we have to
wait to _createPanel until the first show call. Since show() is called
after the right-click sends some data to the panel, it's actually much
simpler to just synchronously create the panel. Changing it to do that
now.