WIP - still debugging
WebChannel connects if the data/default.html page is loaded in a tab,
but not if it's loaded in a frame inside a XUL panel. Why? Dunno.

Pausing to consider just using the SDK panel, removing it from the DOM,
modifying the attributes, then putting it back in the DOM. Reseating XUL
elements causes the XBL bindings to be adjusted, so we might be able to
keep the SDK panel's port but also make the panel draggable by removing
the 'type' attribute and adding the 'backdrag' attribute, as well as
adding the drag handle XUL element.