OK. Next, try relocating panel on window close.
I'm not even sure this is the right user experience, but let's try it.

1) When popupclosing is caught, cancel it and use the window mediator to
look for other windows. If one is found, attach to it. Otherwise, let
the player be destroyed.

If this works, we'll next position the panel absolutely.

If that works, we'll then have the iframe send the addon a signal to
alert that the panel really should be closed.

I kinda wonder if a chromeless window wouldnt' be better though.