Not there yet.
Content worker sends a signal when it loads, and we wait to try to send
anything until that first signal is received.

This doesn't seem to work: still getting some worker errors.

I suspect the panel disappearing bug (after trying to load a panel in a
second window) that we saw a few weeks back is due to this bug:
https://dxr.mozilla.org/mozilla-central/source/addon-sdk/source/lib/sdk/panel.js#114-115

Looks like I have some more reading to do in that file tomorrow, to try
to understand why the workers aren't being set up correctly.

Maybe we need to wait for the worker to be fully shut down before
creating a new panel?