Hmm.
Update: Think I almost have cross-window panels done. Looks like the
panel isn't opening when the window is switched because SDK code isn't
correctly setting the iframe's src. The iframe is never loaded, so the
panel has width and height 0, the code inside the iframe never
loads--hence, the worker's never ready, and throws when you attempt to
contact it. (But then, how is the worker sending an "I'm here" message?
So strange.)

1. track if the old worker actually is detached by listening for the
'detach' message and throwing an Error (to make sure it's seen).

2. try manually calling setURL after the panel is created, just to see
if bumping it a little later would help.