Definitely an SDK bug. Found a workaround.
For some reason, when show is called, at line 291 of sdk/panel.js, this
check returns false:
  if (!isDisposed(this))
    domPanel.show(view, options, anchorView)

And isDisposed comes from sdk/panel/utils.js.

So why would this show code be unable to run? Because the panel's not
ready yet.

Maybe we should check for the panel.isShowing after calling show?
Maybe reseat the panel in the DOM and call show again?
What code modifies the set of views that isDisposed relies on?
Or is the problem that the set is GC'd, not explicitly freed?