Got it. Multiple windows working :neckbeard: :neckbeard: :neckbeard: :neckbeard: :neckbeard:
Waiting for XBL to finish was the crucial thing. I'm not even sure we
need to wait for the SDK to catch up.

----

Try waiting till panel is in XUL DOM and registered with SDK.

Also, have the secondary attach functions get the panel themselves,
rather than pass a reference around. I want to rule out any random
scoping issue.