moz-stack doesn't work.
Looks like the panel utils style method is throwing again.

We will reintroduce that wait, but now I see we can check for an element
with the id sdk-panel-style inside the panel; once that appears, we are
good to mess with the XUL attributes.

so, in the cascading listener, poll for sdk-panel-style, *then* make
draggable. here we go.