OK, draggable on win/mac, topmost on win/mac/linux.
Just need to figure out a suitable positioning system on linux.

I don't want to track the location of the window just for this one
platform.

I also don't want to toggle the titlebar on hover just for linux. I
haven't gotten that to work yet.

What if we just call whatever window manager method, via js-ctypes, that
would be called by the linux-specific `beginWindowMove` method?
Trying that next.